--- src/main/java/dk/sdu/se_f22/BrandModule/Management/Bim.java
@@ -1,5 +1,0 @@
-package dk.sdu.se_f22.BrandModule.Management;
-
-public class Bim implements IBim {
-
-}
