--- src/main/java/dk/sdu/se_f22/BrandModule/Management/json/brands.json
@@ -1,114 +1,0 @@
-{
-    "Acer": {
-        "description": "Acer Inc. is a Taiwanese multinational hardware and electronics corporation specializing in advanced electronics technology, headquartered in Xizhi, New Taipei City. Its products include desktop PCs, laptop PCs (clamshells, 2-in-1s, convertibles and Chromebooks), tablets, servers, storage devices, virtual reality devices, displays, smartphones and peripherals, as well as gaming PCs and accessories under its Predator brand. Acer is the world's 6th-largest PC vendor by unit sales as of January 2021.",
-        "founded": "1 August 1976",
-        "headquarters": "Xizhi, New Taipei, Taiwan",
-        "products": [
-            "personal computers",
-            "smartphones",
-            "servers",
-            "storage",
-            "handhelds",
-            "monitors",
-            "LEDs",
-            "LCDs",
-            "video projectors",
-            "e-business"
-        ]
-    },
-    "HP": {
-        "description": "HP Inc. is an American multinational information technology company headquartered in Palo Alto, California, that develops personal computers (PCs), printers and related supplies, as well as 3D printing solutions.",
-        "founded": "January 1, 1939",
-        "headquarters": "Palo Alto, California, United States",
-        "products": [
-            "Personal computers",
-            "printers",
-            "digital press",
-            "3D printers",
-            "scanners",
-            "copiers",
-            "monitors"
-        ]
-    },
-    "Dell": {
-        "description": "Dell is an American company that develops, sells, repairs, and supports computers and related products and services, and is owned by its parent company of Dell Technologies. Founded in 1984 by Michael Dell, the company is one of the largest technology corporations in the world, employing more than 165,000 people in the United States (US) and around the world. Dell sells personal computers (PCs), servers, data storage devices, network switches, software, computer peripherals, HDTVs, cameras, printers, and electronics built by other manufacturers. The company is well known for its innovations in supply chain management and electronic commerce, particularly its direct-sales model and its \"build-to-order\" or \"configure to order\" approach to manufacturing—delivering individual PCs configured to customer specifications. Dell was a pure hardware vendor for much of its existence, but with the acquisition in 2009 of Perot Systems, Dell entered the market for IT services. The company has since made additional acquisitions in storage and networking systems, with the aim of expanding their portfolio from offering computers only to delivering complete solutions for enterprise customers.",
-        "founded": "February 1, 1984",
-        "headquarters": "Round Rock, Texas, US",
-        "products": [
-            "Personal computers",
-            "Servers",
-            "Peripherals",
-            "Smartphones",
-            "Televisions"
-        ]
-    },
-    "Samsung": {
-        "description": "The Samsung Group (or simply Samsung, stylized in logo as SΛMSUNG) is a South Korean multinational manufacturing conglomerate headquartered in Samsung Town, Seoul, South Korea. It comprises numerous affiliated businesses, most of them united under the Samsung brand, and is the largest South Korean chaebol (business conglomerate). As of 2020, Samsung has the 8th highest global brand value.",
-        "founded": "1 March 1938",
-        "headquarters": "Seocho District, Seoul, South Korea",
-        "products": [
-            "Clothing",
-            "automotive",
-            "chemicals",
-            "consumer electronics",
-            "electronic components",
-            "medical equipment",
-            "semiconductors",
-            "solid state drives",
-            "DRAM",
-            "flash memory",
-            "ships",
-            "telecommunications equipment",
-            "home appliance"
-        ]
-    },
-    "Intel": {
-        "description": "Intel Corporation, stylized as intel, is an American multinational corporation and technology company headquartered in Santa Clara, California. It is the world's largest semiconductor chip manufacturer by revenue, and is the developer of the x86 series of microprocessors, the processors found in most personal computers (PCs). Incorporated in Delaware, Intel ranked No. 45 in the 2020 Fortune 500 list of the largest United States corporations by total revenue for nearly a decade, from 2007 to 2016 fiscal years. Intel supplies microprocessors for computer system manufacturers such as Acer, Lenovo, HP, and Dell. Intel also manufactures motherboard chipsets, network interface controllers and integrated circuits, flash memory, graphics chips, embedded processors and other devices related to communications and computing.",
-        "founded": "July 18, 1968",
-        "headquarters": "San Francisco California, U.S.",
-        "products": [
-            "Central processing units",
-            "Microprocessors",
-            "Integrated graphics processing units (iGPU)",
-            "Systems-on-chip (SoCs)",
-            "Motherboard chipsets",
-            "Network interface controllers",
-            "Modems",
-            "Mobile phones",
-            "Solid state drives",
-            "Wi-Fi and Bluetooth Chipsets",
-            "Flash memory",
-            "Vehicle automation sensors"
-        ]
-    },
-    "Lenovo": {
-        "description": "Lenovo Group Limited, often shortened to Lenovo, is a Chinese / American multinational technology company specializing in designing, manufacturing, and marketing consumer electronics, personal computers, software, business solutions, and related services. Products manufactured by the company include desktop computers, laptops, tablet computers, smartphones, workstations, servers, supercomputers, electronic storage devices, IT management software, and smart televisions. Its best-known brands include IBM's ThinkPad business line of laptop computers, the IdeaPad, Yoga, and Legion consumer lines of laptop computers, and the IdeaCentre and ThinkCentre lines of desktop computers. As of January 2021, Lenovo is the world's largest personal computer vendor by unit sales.",
-        "founded": "1 November 1984",
-        "headquarters": "Hong Kong",
-        "products": [
-            "Personal computers",
-            "smartphones",
-            "servers",
-            "supercomputers",
-            "peripherals",
-            "printers",
-            "televisions",
-            "scanners",
-            "storage devices"
-        ]
-    },
-    "MSI": {
-        "description": "Micro-Star International Co., Ltd is a Taiwanese multinational information technology corporation headquartered in New Taipei City, Taiwan. It designs, develops and provides computer hardware, related products and services, including laptops, desktops, motherboards, graphics cards, All-in-One PCs, servers, industrial computers, PC peripherals, car infotainment products, etc.",
-        "founded": "4 August 1986",
-        "headquarters": "Zhonghe, New Taipei, Taiwan",
-        "products": [
-            "Personal computers",
-            "Motherboards",
-            "Graphics cards",
-            "PC components",
-            "Peripherals",
-            "Monitors",
-            "Server hardware"
-        ]
-    }
-}
