--- src/main/java/dk/sdu/se_f22/BrandModule/Management/services/IIndexingService.java
@@ -1,4 +1,0 @@
-package dk.sdu.se_f22.BrandModule.Management.services;
-
-public interface IIndexingService {
-}
