--- src/main/java/dk/sdu/se_f22/BrandModule/Management/services/IndexingService.java
@@ -1,5 +1,0 @@
-package dk.sdu.se_f22.BrandModule.Management.services;
-
-public class IndexingService implements IIndexingService {
-
-}
